refactor: move to format directory
