+++ src/read.rs
@@ -11,9 +11,10 @@ use crate::read::readers::{ZipFileReader, ZipFileSeekReader};
 use crate::result::{ZipError, ZipResult, invalid};
 use crate::spec::{
     CentralDirectoryEndInfo, DataAndPosition, FixedSizeBlock, ZIP64_BYTES_THR,
-    ZipCentralEntryBlock, ZipFlags,
+    ZipCentralEntryBlock
 };
-use crate::types::{SimpleFileOptions, System, ZipFileData, ffi};
+use crate::format::flags::{ZipFlags, System};
+use crate::types::{SimpleFileOptions, ZipFileData, ffi};
 use crate::unstable::LittleEndianReadExt;
 use core::mem::replace;
 use indexmap::IndexMap;
