+++ src/types.rs
@@ -10,18 +10,20 @@ use crate::result::{ZipError, ZipResult, invalid};
 use crate::spec::is_dir;
 use crate::spec::{
     self, FixedSizeBlock, Magic, Zip64DataDescriptorBlock, ZipCentralEntryBlock,
-    ZipDataDescriptorBlock, ZipFlags, ZipLocalEntryBlock,
+    ZipDataDescriptorBlock, ZipLocalEntryBlock,
 };
+use crate::format::flags::ZipFlags;
 use crate::write::FileOptionExtension;
 use crate::zipcrypto::ZipCryptoKeys;
-use core::fmt::Debug;
-use core::fmt::Display;
 use core::marker::PhantomData;
 use std::borrow::Cow;
 use std::ffi::OsStr;
 use std::io::{Read, Seek, SeekFrom, Take};
 use std::path::{Path, PathBuf};
 use std::sync::{Arc, OnceLock};
+
+pub(crate) use crate::format::aes::{AesVendorVersion, AesMode};
+pub(crate) use crate::format::flags::System;
 
 pub(crate) mod ffi {
     pub const S_IFDIR: u32 = 0o0_040_000;
@@ -35,107 +37,6 @@ pub(crate) struct ZipRawValues {
     pub(crate) uncompressed_size: u64,
 }
 
-/// System inside `version made by` (upper byte)
-/// Reference: 4.4.2.2
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
-#[allow(clippy::upper_case_acronyms)]
-#[repr(u8)]
-pub enum System {
-    /// `MS-DOS` and `OS/2` (`FAT` / `VFAT` / `FAT32` file systems; default on Windows)
-    Dos = 0,
-    /// `Amiga`
-    Amiga = 1,
-    /// `OpenVMS`
-    OpenVMS = 2,
-    /// Default on Unix; default for symlinks on all platforms
-    Unix = 3,
-    /// `VM/CMS`
-    VmCms = 4,
-    /// `Atari ST`
-    AtariSt = 5,
-    /// `OS/2 H.P.F.S.`
-    Os2 = 6,
-    /// Legacy `Mac OS`, pre `OS X`
-    Macintosh = 7,
-    /// `Z-System`
-    ZSystem = 8,
-    /// `CP/M`
-    CPM = 9,
-    /// Windows NTFS (with extra attributes; not used by default)
-    WindowsNTFS = 10,
-    /// `MVS (OS/390 - Z/OS)`
-    MVS = 11,
-    /// `VSE`
-    VSE = 12,
-    /// `Acorn Risc`
-    AcornRisc = 13,
-    /// `VFAT`
-    VFAT = 14,
-    /// alternate MVS
-    AlternateMVS = 15,
-    /// `BeOS`
-    BeOS = 16,
-    /// `Tandem`
-    Tandem = 17,
-    /// `OS/400`
-    Os400 = 18,
-    /// `OS X` (Darwin) (with extra attributes; not used by default)
-    OsDarwin = 19,
-    /// unused
-    #[default]
-    Unknown = 255,
-}
-
-impl System {
-    /// Parse `version_made_by` block in local entry block.
-    #[must_use]
-    pub fn from_version_made_by(version_made_by: u16) -> Self {
-        // Extract upper byte from little-endian representation
-        let upper_byte = version_made_by.to_le_bytes()[1];
-        System::from(upper_byte) // from u8
-    }
-
-    /// Extract the system and version from a `version_made_by` field.
-    /// The first byte (lower) is the version, and the second byte (upper) is the system.
-    pub(crate) fn extract_bytes(version_made_by: u16) -> (u8, Self) {
-        let bytes = version_made_by.to_le_bytes();
-        (bytes[0], Self::from(bytes[1]))
-    }
-}
-
-impl From<u8> for System {
-    fn from(system: u8) -> Self {
-        match system {
-            0 => System::Dos,
-            1 => System::Amiga,
-            2 => System::OpenVMS,
-            3 => System::Unix,
-            4 => System::VmCms,
-            5 => System::AtariSt,
-            6 => System::Os2,
-            7 => System::Macintosh,
-            8 => System::ZSystem,
-            9 => System::CPM,
-            10 => System::WindowsNTFS,
-            11 => System::MVS,
-            12 => System::VSE,
-            13 => System::AcornRisc,
-            14 => System::VFAT,
-            15 => System::AlternateMVS,
-            16 => System::BeOS,
-            17 => System::Tandem,
-            18 => System::Os400,
-            19 => System::OsDarwin,
-            _ => System::Unknown,
-        }
-    }
-}
-
-impl From<System> for u8 {
-    fn from(system: System) -> Self {
-        system as u8
-    }
-}
 
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub(crate) enum EncryptWith<'k> {
@@ -766,119 +667,6 @@ impl ZipFileData {
     }
 }
 
-/// The encryption specification used to encrypt a file with AES.
-///
-/// According to the [specification](https://www.winzip.com/win/en/aes_info.html#winzip11) AE-2
-/// does not make use of the CRC check.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-#[repr(u16)]
-pub enum AesVendorVersion {
-    Ae1 = 0x0001,
-    Ae2 = 0x0002,
-}
-
-impl AesVendorVersion {
-    /// As u16
-    #[must_use]
-    pub const fn as_u16(self) -> u16 {
-        self as u16
-    }
-
-    /// Returns `true` if the data is encrypted using AE2.
-    #[cfg(feature = "aes-crypto")]
-    pub const fn is_ae2_encrypted(&self) -> bool {
-        matches!(self, AesVendorVersion::Ae2)
-    }
-
-    /// `false` since the feature `aes-crypto` is not enabled
-    #[cfg(not(feature = "aes-crypto"))]
-    pub const fn is_ae2_encrypted(&self) -> bool {
-        false
-    }
-}
-
-impl TryFrom<u16> for AesVendorVersion {
-    type Error = &'static str;
-
-    fn try_from(value: u16) -> Result<Self, Self::Error> {
-        let aes_vendor_version = match value {
-            0x0001 => AesVendorVersion::Ae1,
-            0x0002 => AesVendorVersion::Ae2,
-            _ => return Err("Invalid AES vendor version"),
-        };
-        Ok(aes_vendor_version)
-    }
-}
-
-impl From<AesVendorVersion> for u16 {
-    fn from(value: AesVendorVersion) -> Self {
-        value as u16
-    }
-}
-
-/// AES variant used.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-#[cfg_attr(feature = "_arbitrary", derive(arbitrary::Arbitrary))]
-#[repr(u8)]
-pub enum AesMode {
-    /// 128-bit AES encryption.
-    Aes128 = 0x01,
-    /// 192-bit AES encryption.
-    Aes192 = 0x02,
-    /// 256-bit AES encryption.
-    Aes256 = 0x03,
-}
-
-impl AesMode {
-    /// As u8
-    #[must_use]
-    pub const fn as_u8(self) -> u8 {
-        self as u8
-    }
-}
-
-impl Display for AesMode {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        match self {
-            Self::Aes128 => write!(f, "AES-128"),
-            Self::Aes192 => write!(f, "AES-192"),
-            Self::Aes256 => write!(f, "AES-256"),
-        }
-    }
-}
-
-impl TryFrom<u8> for AesMode {
-    type Error = &'static str;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        let mode = match value {
-            0x01 => AesMode::Aes128,
-            0x02 => AesMode::Aes192,
-            0x03 => AesMode::Aes256,
-            _ => return Err("Invalid AES encryption strength"),
-        };
-        Ok(mode)
-    }
-}
-
-#[cfg(feature = "aes-crypto")]
-impl AesMode {
-    /// Length of the salt for the given AES mode.
-    #[must_use]
-    pub const fn salt_length(&self) -> usize {
-        self.key_length() / 2
-    }
-
-    /// Length of the key for the given AES mode.
-    #[must_use]
-    pub const fn key_length(&self) -> usize {
-        match self {
-            Self::Aes128 => 16,
-            Self::Aes192 => 24,
-            Self::Aes256 => 32,
-        }
-    }
-}
 
 #[cfg(test)]
 mod tests {
